LSO: include check if lso abstract or extro image exists. If not do not execute deletion of file(s). (#37590) (#6234)

+++ Modules/LearningSequence/classes/Settings/class.ilLearningSequenceFilesystem.php
@@ -1,7 +1,5 @@
 <?php
 
-declare(strict_types=1);
-
 /**
  * This file is part of ILIAS, a powerful learning management system
  * published by ILIAS open source e-Learning e.V.
@@ -17,6 +15,8 @@ declare(strict_types=1);
  * https://github.com/ILIAS-eLearning
  *
  *********************************************************************/
+
+declare(strict_types=1);
 
 /**
  * Storage of images in settings.
@@ -69,7 +69,9 @@ class ilLearningSequenceFilesystem extends ilFileSystemAbstractionStorage
             $delete = $settings->getExtroImage();
             $settings = $settings->withExtroImage();
         }
-        $this->deleteFile($delete);
+        if (!empty($delete)) {
+            $this->deleteFile($delete);
+        }
         return $settings;
     }
